Fix editor crash when deleting forwards at EOL
This commit fixes a crash caused by an off-by-one error. When deleting
forwards in an edit buffer with the cursor at the end, the delete count
would not be appropriately limited due to an off-by-one error in the
bounds check.

In addition to fixing the issue, a test to catch it was also added.