Fix incorrect macro expansion through variable
Previously, the following input would produce incorrect results:

    (def x '(for i (1 2 3) (print i)))
    x

This would actually *execute* the contained form, when the correct
behavior would be to display the quoted form without executing it. This
was happening because the planning engine was dereferencing (i.e.
evaluating) isolated symbols, so when the form was executed it would be
evaluated twice instead of once, as it should be.

This commit fixes the issue by not auto-dereferencing variables inside
expression elements.